refactor: 💡 move integration to a separate file
